refactor(models): 디렉터리 네이밍 변경
models에서 domain으로 네이밍 변경
