+++ src/domain/RacingWinnerRecorder/RacingWinnerRecorder.test.js
