+++ src/domain/RandomMoveCountMaker/RandomMoveCountMaker.js
