+++ src/domain/RandomMoveCountMaker/RandomMoveCountMaker.test.js
