+++ src/domain/RandomMoveCountMaker/constant.js
